updating workspace service and commenting

+++ todo/src/main/java/com/todo/service/WorkspaceService.java
@@ -6,6 +6,8 @@ import com.todo.model.Workspace;
 import com.todo.repository.LaneRepository;
 import com.todo.repository.WorkspaceRepository;
 import io.swagger.v3.oas.annotations.Hidden;
+import io.swagger.v3.oas.annotations.Parameter;
+import org.hibernate.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -30,8 +32,8 @@ public class WorkspaceService {
 
     /**
      * Get a workspace by id.
-     * @param id
-     * @return
+     * @param id The id of the workspace to be retrieved.
+     * @return The workspace with the given id.
      */
     public Workspace getWorkspaceById(int id){
         return workspaceRepository.findById(id).orElse(null);
@@ -39,8 +41,8 @@ public class WorkspaceService {
 
     /**
      * Create a workspace.
-     * @param name
-     * @return
+     * @param name The name of the workspace.
+     * @return The created workspace.
      */
     public Workspace createWorkspace(String name){
         Workspace workspace = new Workspace();
@@ -51,17 +53,32 @@ public class WorkspaceService {
 
     /**
      * Delete a workspace.
-     * @param workspace
+     * @param workspaceId The id of the workspace to be deleted.
      */
-    public void deleteWorkspace(Workspace workspace){
+    public void deleteWorkspace(int workspaceId) throws ObjectNotFoundException{
+        Workspace workspace = workspaceRepository.findById(workspaceId).orElse(null);
+        if(workspace == null) throw new ObjectNotFoundException(workspaceId, "Workspace");
         workspaceRepository.delete(workspace);
     }
 
+    /**
+     * Get all lanes in a workspace.
+     * @param workspace The workspace to get lanes from.
+     * @return A list of all lanes in the workspace.
+     */
     public List<Lane> getAllLanesInWorkspace(Workspace workspace){
         return workspace.getLanes();
     }
 
-    public Lane createLane(String name, Workspace workspace){
+    /**
+     * Create a lane in a workspace.
+     * @param name The name of the lane.
+     * @param workspaceId The id of the workspace to create the lane in.
+     * @return The created lane.
+     */
+    public Lane createLane(String name, int workspaceId){
+        Workspace workspace = workspaceRepository.findById(workspaceId).orElse(null);
+        if(workspace == null) throw new ObjectNotFoundException(workspaceId, "Workspace");
         Lane lane = new Lane();
         lane.setName(name);
         lane.setWorkspace(workspace);
@@ -73,6 +90,11 @@ public class WorkspaceService {
         return lane;
     }
 
+    /**
+     * Delete a lane.
+     * @param lane The lane to be deleted.
+     * @return The workspace the lane was deleted from.
+     */
     public Workspace deleteLane(Lane lane){
         Workspace workspace = lane.getWorkspace();
         workspace.getLanes().remove(lane);
@@ -82,6 +104,11 @@ public class WorkspaceService {
         return workspace;
     }
 
+    /**
+     * Get a lane by id.
+     * @param laneId The id of the lane to be retrieved.
+     * @return The lane with the given id.
+     */
     public WorkspaceDTO moveLane(int laneId, int position){
 
         Lane lane = laneRepository.findById(laneId).orElse(null);
@@ -100,11 +127,20 @@ public class WorkspaceService {
         return new WorkspaceDTO(workspace);
     }
 
+    /**
+     * Update a workspace.
+     * @param workspace The workspace to be updated.
+     * @return The updated workspace.
+     */
     public Workspace update(Workspace workspace) {
         workspaceRepository.save(workspace);
         return workspace;
     }
 
+    /**
+     * Get all workspaces.
+     * @return A list of all workspaces.
+     */
     public List<WorkspaceDTO> getAllWorkspaces() {
         List<Workspace> workspaces = workspaceRepository.findAll();
         return workspaces.stream()
@@ -112,6 +148,10 @@ public class WorkspaceService {
                 .collect(java.util.stream.Collectors.toList());
     }
 
+    /**
+     * Update the positions of all lanes in a workspace.
+     * @param workspace The workspace to update the lane positions in.
+     */
     private void updateLanePositions(Workspace workspace) {
         for(int i = 0; i < workspace.getLanes().size(); i++){
             workspace.getLanes().get(i).setPosition(i);
